feat: change category to int

+++ src/rutcovschi/intercon-be/Intercon.Application/Extensions/Mappers/BusinessMapper.cs
@@ -29,6 +29,6 @@ public static class BusinessMapper
             business.LogoId is not null ? business.Logo?.Path : null,
             business.Address,
             business.ReviewsCount,
-            business.Category);
+            (int)business.Category);
     }
 }
